create seperate test file and added api into endpoints.json
